Renamed shared folder from interfaces to models
